llm_model/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pymongo==4.6.0
pinecone-client==2.2.4
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2 